NamedTuple Step, set_state for wrappers

+++ src/marlenv/models/spaces.py
@@ -112,7 +112,7 @@ class ActionSpace(Space, Generic[S]):
             else:
                 m = None
             res.append(self.individual_action_space.sample(m))
-        return np.array(res, dtype=np.int64)
+        return np.array(res, dtype=np.int32)
 
 
 class DiscreteActionSpace(ActionSpace[DiscreteSpace]):
